Remove indentation sensitivity
Instead of layout, we take a page out of lua's approach and forgo any explicit
statement separators all together. Flora's syntax is simple enough that this
is possible (with the only major change being a few additional ',' in match/handle
expressions) and this allows auto-formatting foreign files that embed flora expressions.